requirements.txt
python-dotenv
pandas
numpy
matplotlib
seaborn
pymongo
certifi
scikit-learn
dill
pyaml


#-e . 